move content from unused test to existing one with same name
