+++ tests/AppBundle/Entity/AdmissionPeriodEntityUnitTest.php
@@ -5,6 +5,7 @@ namespace Tests\AppBundle\Entity;
 
 use AppBundle\Entity\AdmissionPeriod;
 use AppBundle\Entity\InfoMeeting;
+use AppBundle\Entity\Department;
 use DateTime;
 use PHPUnit\Framework\TestCase;
 
@@ -51,5 +52,55 @@ class AdmissionPeriodEntityUnitTest extends TestCase
         $admissionPeriod->getInfoMeeting()->setDate((new DateTime())->modify('-1 hour'));
         $admissionPeriod->getInfoMeeting()->setShowOnPage(true);
         $this->assertFalse($admissionPeriod->shouldSendInfoMeetingNotifications());
+    }
+
+    // Check whether the setstartDate function is working correctly
+    public function testSetstartDate()
+    {
+
+        // New datetime variable
+        $today = new DateTime('now');
+
+        // new entity
+        $admissionPeriod = new AdmissionPeriod();
+
+        // Use the setstartDate method
+        $admissionPeriod->setStartDate($today);
+
+        // Assert the result
+        $this->assertEquals($today, $admissionPeriod->getStartDate());
+    }
+
+    // Check whether the setendDate function is working correctly
+    public function testSetendDate()
+    {
+
+        // New datetime variable
+        $today = new DateTime('now');
+
+        // new entity
+        $admissionPeriod = new AdmissionPeriod();
+
+        // Use the setendDate method
+        $admissionPeriod->setEndDate($today);
+
+        // Assert the result
+        $this->assertEquals($today, $admissionPeriod->getEndDate());
+    }
+
+    // Check whether the setDepartment function is working correctly
+    public function testSetDepartment()
+    {
+        // new entity
+        $admissionPeriod = new AdmissionPeriod();
+
+        // Dummy entity
+        $department = new Department();
+
+        // Use the setDepartment method
+        $admissionPeriod->setDepartment($department);
+
+        // Assert the result
+        $this->assertEquals($department, $admissionPeriod->getDepartment());
     }
 }
